requirements.txt
pandas
openpyxl
groq
python-dotenv
numpy
certifi
streamlit
yfinance
pip-system-certs
pyxirr
scikit-learn